pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "desktop-voice-assistant"
version = "0.1.0"
description = "Cross-platform desktop voice assistant with Tkinter GUI, STT/TTS, and modular skills"
authors = [
{ name = "Mobin Yousefi", email = "mobin@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
"SpeechRecognition>=3.10.0",
"pyttsx3>=2.90",
"wikipedia>=1.4.0",
"requests>=2.32.0",
"beautifulsoup4>=4.12.0",
"wolframalpha>=5.0.0; python_version<'3.13'",
"pyjokes>=0.6.0",
"python-dotenv>=1.0.1",
]


[project.optional-dependencies]
win = ["winshell>=0.6"]


[project.scripts]
voice-assistant = "voice_assistant.app:main"


[tool.setuptools]
package-dir = {"" = "src"}


[tool.setuptools.packages.find]
where = ["src"]
include = ["voice_assistant*"]


[tool.black]
line-length = 100


[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "W"]
fix = true


[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]